improv: Fetch purposes when starting, and show in UI
The way it is shown is not ideal, but we want to update the design here
anyway. So I didn't try to style it especially well.

This fetches purposes from the server when used in Gnome Control Center
(falling back to a version hard-coded in hp-vendor), but doesn't attempt
fetching in Gnome Initial Setup. This is apparently the desired
behavior, though it's not great that it takes some time to communicate
with the server for this, before it's possible to show purpose text or
send opt in.