Add data sample json file
This was generated by clearing `/var/hp-vendor`, and running `hp-vendor
consent`, `hp-vendor daily`, and `hp-vendor daily-upload`.

Replaced UUIDs with "ffffffff-ffff-ffff-ffff-ffffffffffff", and device
serial number with "0000000000". This is important since these allow
authenticating with the API.

For good measure, replaced `serial_number` of RAM, nvme drives, etc.
with "ABCDE".

The format of the data is a bit funny since it is just exactly a list of
uploads to the server, but that's what the server returns. It also is
limited to data currently collected, while some things may still need to
be added later.